+++ codex-rs/sandboxing/BUILD.bazel
@@ -1,0 +1,6 @@
+load("//:defs.bzl", "codex_rust_crate")
+
+codex_rust_crate(
+    name = "sandboxing",
+    crate_name = "codex_sandboxing",
+)
